[30197] Use separate ENV to trigger frontend angular recompilation
This PR introduces two env variables for packaged installations:

- RECOMPILE_RAILS_ASSETS will be set when a server prefix set and
backend assets need to be recompiled or when a custom gemfile is present

- RECOMPILE_ANGULAR_ASSETS will be undefined by default and will result
in skipping angular recompilation if `RECOMPILE_RAILS_ASSETS=true`. Can
be manually set by the user if frontend recompilation is needed (due to
a custom frontend plugin).

Also checks if the server prefix changed from any previous value, and
forces recompilation in that case as well

https://community.openproject.com/wp/30197